Rename bitcoin-move light_client.move to bitcoin.move
