+++ crates/rooch-executor/src/actor/executor.rs
@@ -23,7 +23,7 @@ use moveos_types::transaction::{FunctionCall, MoveOSTransaction, VerifiedMoveAct
 use rooch_genesis::FrameworksGasParameters;
 use rooch_store::RoochStore;
 use rooch_types::address::MultiChainAddress;
-use rooch_types::bitcoin::light_client::BitcoinLightClientModule;
+use rooch_types::bitcoin::BitcoinModule;
 use rooch_types::framework::address_mapping::AddressMappingModule;
 use rooch_types::framework::auth_validator::{AuthValidatorCaller, TxValidateResult};
 use rooch_types::framework::ethereum_light_client::EthereumLightClientModule;
@@ -130,7 +130,7 @@ impl ExecutorActor {
         match RoochMultiChainID::try_from(chain_id.id())? {
             RoochMultiChainID::Bitcoin => {
                 let action = VerifiedMoveAction::Function {
-                    call: BitcoinLightClientModule::create_submit_new_block_call_bytes(
+                    call: BitcoinModule::create_submit_new_block_call_bytes(
                         block_height,
                         block_hash,
                         block_body,
