+++ crates/rooch-framework-tests/src/tests/mod.rs
@@ -1,7 +1,7 @@
 // Copyright (c) RoochNetwork
 // SPDX-License-Identifier: Apache-2.0
 
-mod bitcoin_light_client_test;
+mod bitcoin_test;
 mod brc20_test;
 mod chain_id_test;
 mod empty_tests;
