+++ frameworks/bitcoin-move/doc/genesis.md
@@ -12,8 +12,8 @@
 <pre><code><b>use</b> <a href="">0x1::option</a>;
 <b>use</b> <a href="">0x2::signer</a>;
 <b>use</b> <a href="">0x2::tx_context</a>;
+<b>use</b> <a href="light_client.md#0x4_bitcoin">0x4::bitcoin</a>;
 <b>use</b> <a href="brc20.md#0x4_brc20">0x4::brc20</a>;
-<b>use</b> <a href="light_client.md#0x4_light_client">0x4::light_client</a>;
 <b>use</b> <a href="network.md#0x4_network">0x4::network</a>;
 <b>use</b> <a href="ord.md#0x4_ord">0x4::ord</a>;
 <b>use</b> <a href="utxo.md#0x4_utxo">0x4::utxo</a>;
